+++ tools.go
@@ -3,16 +3,19 @@
 // license that can be found in the LICENSE file.
 
 //go:build tools
-// +build tools
 
 package tools
 
 import (
-	_ "github.com/actgardner/gogen-avro/v9/cmd/gogen-avro"
+	_ "github.com/golangci/golangci-lint/v2/cmd/golangci-lint"
 	_ "github.com/google/addlicense"
+	_ "github.com/google/osv-scanner/v2/cmd/osv-scanner"
 	_ "github.com/incu6us/goimports-reviser/v3"
 	_ "github.com/kisielk/errcheck"
 	_ "github.com/maxbrunsfeld/counterfeiter/v6"
-	_ "golang.org/x/lint/golint"
+	_ "github.com/onsi/ginkgo/v2/ginkgo"
+	_ "github.com/securego/gosec/v2/cmd/gosec"
+	_ "github.com/segmentio/golines"
+	_ "github.com/shoenig/go-modtool"
 	_ "golang.org/x/vuln/cmd/govulncheck"
 )